requirements.txt
ultralytics
opencv-python-headless
gradio
numpy
pillow